// eshop-web/src/main/resources/client/src/app/analytic-logging/enums/ga-search-category.enums.ts
export enum GA_SEARCH_CATEGORIES {
    LICENSE_PLATE = 'license_plate',
    VEHICLE_NUMBER = 'vehicle_number',
    VEHICLE_DESCRIPTION = 'vehicle_description',
    VIN = 'vin',
    ARTICLE_NUMBER = 'article_number',
    ARTICLE_DESCRIPTION = 'article_description',
    FREE_TEXT = 'free_text',
    TYRE_DIMENSIONS = 'tyre_dimensions',
    TYRE_SUPPLIER = 'tyre_supplier'
}

export enum GA_USER_TYPE {
    CUSTOMER = 'Customer',
    SALES = 'Sales'
}