// eshop-web/src/main/resources/client/src/app/analytic-logging/models/ga-product-field.model.ts
export class GaProductField {
    item_id: string;
    item_name: string;
    item_brand: string;
    currency: string;
    price: number;
    quantity: number;
    item_category: string;
    item_category2: string;
    item_category3: string;
    item_category4: string;
    item_category5: string;
    item_availability: string;
    item_list_name?: string;
    item_sub_list_name?: string;
    item_availability_partner?: string;
    source_id?: string;
    source_description?: string;
}